auth/requirements.txt
astroid==2.15.5
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dill==0.3.6
Flask==2.3.2
Flask-MySQL==1.5.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jwt==1.3.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
parso==0.8.3
platformdirs==3.8.0
protobuf==3.20.3
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
PyMySQL==1.1.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
Werkzeug==2.3.6
wrapt==1.15.0
